pyproject.toml
[tool.poetry]
name = "repl_python3_autonomous-study-discord-bot"
version = "0.1.0"
description = ""
authors = ["TomasAmaro <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
